refactor : 검증에 대한 유틸 클래스의 패키지 이동
(cherry picked from commit f752ec107311df5d48d8c4c9cf6be67b96d1c74c)
