+++ src/main/java/view/InputView.java
@@ -3,8 +3,8 @@ package view;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
-import view.validation.DigitValidator;
-import view.validation.SplitValidator;
+import util.DigitValidator;
+import util.SplitValidator;
 
 public class InputView {
     private static final String CAR_NAMES_DELIMITER = ",";
